Combined Intake.java and Feed.java
